+++ src/legacy/core_plugins/data/public/search/aggs/buckets/date_range.ts
@@ -20,20 +20,18 @@ import { get } from 'lodash';
 import moment from 'moment-timezone';
 import { i18n } from '@kbn/i18n';
 import { npStart } from 'ui/new_platform';
+import { convertDateRangeToString, DateRangeKey } from './lib/date_range';
 import { BUCKET_TYPES } from './bucket_agg_types';
 import { BucketAggType, IBucketAggConfig } from './_bucket_agg_type';
 import { createFilterDateRange } from './create_filter/date_range';
 
 import { KBN_FIELD_TYPES, fieldFormats } from '../../../../../../../plugins/data/public';
 
+export { convertDateRangeToString, DateRangeKey };
+
 const dateRangeTitle = i18n.translate('data.search.aggs.buckets.dateRangeTitle', {
   defaultMessage: 'Date Range',
 });
-
-export interface DateRangeKey {
-  from: number;
-  to: number;
-}
 
 export const dateRangeBucketAgg = new BucketAggType({
   name: BUCKET_TYPES.DATE_RANGE,
@@ -106,16 +104,3 @@ export const dateRangeBucketAgg = new BucketAggType({
     },
   ],
 });
-
-export const convertDateRangeToString = (
-  { from, to }: DateRangeKey,
-  format: (val: any) => string
-) => {
-  if (!from) {
-    return 'Before ' + format(to);
-  } else if (!to) {
-    return 'After ' + format(from);
-  } else {
-    return format(from) + ' to ' + format(to);
-  }
-};
