+++ src/legacy/core_plugins/data/public/search/aggs/buckets/lib/date_range.ts
@@ -17,11 +17,20 @@
  * under the License.
  */
 
-import { SerializedFieldFormat } from '../../../../expressions/common/types';
-import { IFieldFormat } from '../index';
+export interface DateRangeKey {
+  from: number;
+  to: number;
+}
 
-export { asPrettyString } from './as_pretty_string';
-export { getHighlightHtml, getHighlightRequest } from './highlight';
-export { serializeFieldFormat } from './serialize';
-
-export type FormatFactory = (mapping?: SerializedFieldFormat) => IFieldFormat;
+export const convertDateRangeToString = (
+  { from, to }: DateRangeKey,
+  format: (val: any) => string
+) => {
+  if (!from) {
+    return 'Before ' + format(to);
+  } else if (!to) {
+    return 'After ' + format(from);
+  } else {
+    return format(from) + ' to ' + format(to);
+  }
+};
