+++ src/legacy/core_plugins/vis_type_vislib/public/legacy_imports.ts
@@ -18,10 +18,11 @@
  */
 
 export { AggType, AggGroupNames, IAggConfig, IAggType, Schemas } from 'ui/agg_types';
-export { getFormat, getTableAggs } from 'ui/visualize/loader/pipeline_helpers/utilities';
+export { getFormat } from 'ui/visualize/loader/pipeline_helpers/utilities';
 // @ts-ignore
 export { tabifyAggResponse } from 'ui/agg_response/tabify';
 // @ts-ignore
 export { buildHierarchicalData } from 'ui/agg_response/hierarchical/build_hierarchical_data';
 // @ts-ignore
 export { buildPointSeriesData } from 'ui/agg_response/point_series/point_series';
+export { tabifyGetColumns } from '../../../ui/public/agg_response/tabify/_get_columns';
