+++ src/plugins/data/common/field_formats/index.ts
@@ -41,7 +41,7 @@ export {
   TruncateFormat,
 } from './converters';
 
-export { getHighlightRequest } from './utils';
+export { getHighlightRequest, serializeFieldFormat } from './utils';
 
 export { DEFAULT_CONVERTER_COLOR } from './constants/color_default';
 export { FIELD_FORMAT_IDS } from './types';
