+++ src/plugins/data/public/field_formats/field_formats_service.ts
@@ -19,6 +19,8 @@
 
 import { CoreSetup } from 'src/core/public';
 import { FieldFormatsRegistry } from '../../common/field_formats';
+import { deserializeFieldFormat } from './utils/deserialize';
+import { FormatFactory } from '../../common/field_formats/utils';
 
 export class FieldFormatsService {
   private readonly fieldFormatsRegistry: FieldFormatsRegistry = new FieldFormatsRegistry();
@@ -44,6 +46,10 @@ export class FieldFormatsService {
   }
 
   public start() {
+    this.fieldFormatsRegistry.deserialize = deserializeFieldFormat.bind(
+      this.fieldFormatsRegistry as FieldFormatsStart
+    );
+
     return this.fieldFormatsRegistry as FieldFormatsStart;
   }
 }
@@ -52,4 +58,6 @@ export class FieldFormatsService {
 export type FieldFormatsSetup = Pick<FieldFormatsRegistry, 'register'>;
 
 /** @public */
-export type FieldFormatsStart = Omit<FieldFormatsRegistry, 'init' & 'register'>;
+export type FieldFormatsStart = Omit<FieldFormatsRegistry, 'init' & 'register'> & {
+  deserialize: FormatFactory;
+};
