+++ src/test_utils/public/stub_field_formats.ts
@@ -17,13 +17,18 @@
  * under the License.
  */
 import { CoreSetup } from 'kibana/public';
-import { fieldFormats } from '../../plugins/data/public';
+import { DataPublicPluginStart, fieldFormats } from '../../plugins/data/public';
+import { deserializeFieldFormat } from '../../plugins/data/public/field_formats/utils/deserialize';
 
 export const getFieldFormatsRegistry = (core: CoreSetup) => {
   const fieldFormatsRegistry = new fieldFormats.FieldFormatsRegistry();
   const getConfig = core.uiSettings.get.bind(core.uiSettings);
 
   fieldFormatsRegistry.init(getConfig, {});
+
+  fieldFormatsRegistry.deserialize = deserializeFieldFormat.bind(
+    fieldFormatsRegistry as DataPublicPluginStart['fieldFormats']
+  );
 
   return fieldFormatsRegistry;
 };
